PROJECT WEBDEV/requirements.txt
Django==3.2.12
requests==2.28.1
Pillow==10.2.0 